requirements.txt
python-telegram-bot==21.5
kiteconnect==5.0.1
openai==1.40.0
pandas==2.2.2
pandas-ta==0.3.14b
numpy==1.26.4
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.9.5
pytz==2024.1
requests==2.32.3
